+++ src/Http/Resources/UnitResource.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ChrisLorando\LaravelAccurate\Http\Resources;
+
+class UnitResource extends Resource
+{
+    protected string $resourceName = 'unit';
+}